clutter: Change clutter_seat_query_state() API to use sprites
Even though ClutterSeat is not directly involved with ClutterSprite,
and the input thread even less (i.e. access is forbidden), make this
API reliant on ClutterSprite to express the specific pointing focus.

In order to make the API a bit less convoluted, allow a NULL sprite
special case to query for the "pointer" position.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4458>