monitor-manager: Create color devices as part of rebuild
Color devices really are a part of the monitor, even though we don't
store them there (we probably should). When (re)creating monitors, we
also must recreate color devices for them, but can only do so after the
config has been applied because the config determines the properties of
the color device.

Pull the color device recreation from the backend into the
MonitorManagers rebuild process. This will allow other things which
depend on the color device to also work in the rebuild process (such as
the reference white backlight we will add soon).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4014>