tests/dbus-runner: Explicitly disable logind mocking passthrough
The mocked logind by default passes through a session and seat from the
host system. The VKMS and KMS tests are supposed to behave like the
default mode, so they are expected to be the session controller. The
mocked logind is supposed to mock all of this, but not when passthrough
is active. In that case, the real session is taken control over.

Simply tell the mocked logind to not pass through to the host session
when TTY or KVM tests are being used.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4448>