kms/impl-device: Only await flush before mode set
When creating a CRTC frame struct, which handles state related to CRTC
frame scheduling, only await flush if the CRTC is not active. This is
needed to be able to show the hardware cursor before anything ever
changes on primary plane after the initial mode set.

The problem was that on mode set, no CRTC frame had been created, but
after moving the cursor, the frame was created but awaited a KMS update
(empty or not) from the main thread before taking effect (await_flush),
but if nothing changed on screen, such a flush would never arrive.

This is only relevant on e.g. plain mutter or gnome-kiosk with if
application window showing.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4452>