launcher: Take control via method instead of in the constructor
This allows us to create a launcher where we can look up information and
then later decide if we need control. In particular, the
MetaBackendNative needs a non-controlling launcher to figure out if
there is no seat to run headlessly, but in the other case where a seat
exists, it needs to take control.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4448>