keymap/native: Move depressed/latched/locked modifers to abstract class
It'll be used by MetaEisClient, which is not tied to the native backend.
We already have said information in the X11 backend, so just plumb it
there.

This adds a dependency on libxkbcommon on Clutter, even so only for the
types, but this should be harmless, since we depend on it in mutter
already for all backends.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4395>