cogl: Remove pre-condition in Renderer.get_driver_id
As in the winsys set-up, during the renderer_connect "vfunc", we want to
check the set Renderer.driver_id but at that time, the
Renderer.connected is still set to FALSE making this pre-condition fail.

As we want to no longer access the internal fields, let us remove this
condition for now. We can eventually duplicate the driver_id into two
separate things, one is the actual driver_id post connection and one
that is the one to be tried based on the user preference through an env
variable for example.