cogl: Also free custom_winsys_user_data on the Renderer side
Here, we can't just check if the custom_winsys_user_data is set in order
to free it. As in the MetaRendererNative impl, we are passing a
MetaRenderer as the user_data but that is not something we want to free
on our side.

All of this hack-y solutions will be gone as soon as we turn Renderer
into an actual derivable type.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4477>