ci: Update citemplates component version
The image we use now has a non-root user as well.
Gitlab happily defaults to it, but for the time being we
still need to run as root until we fix all the issues.
This was fixed in the template [1] and we need to update
the ref to include the change.

[1] https://gitlab.gnome.org/GNOME/citemplates/-/merge_requests/112

Mirror of https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3735

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4478>