context/main: Fix ifdef for --no-x11 option
When launching mutter the --no-x11 option disables
xwayland, however we were ifdefing it based on if
we built with HAVE_X11. When building with x11=false
the option went missing and mutter always wanted
to launch xwayland.

Use HAVE_XWAYLAND instead (and as expected)
so we can still avoid launching xwayland in a
wayland session.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4485>