monitor: Bias towards smaller scaling factors less strongly
In d03dce4378 ("monitor: Use fractional scale result to decide the
non-fractional scale") a bias for selecting an integer scale from a
fractional scale of 0.25 was introduced. This bias turns out to be a bit
too much. For example the Apple Macbook Retina displays ended up with an
integer scale of 1 whereas users expected a scale of 2.

This commit reduces the bias towards smaller scaling factors slightly,
so the displays in question get the desired scale of 2.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4110
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4490>