mdk: Use full path in desktop file
.desktop files without a valid Exec field are ignored. To be considered
valid, it must be possible to resolve the value to an executable file.

As mutter-devkit is installed in libexec and not in PATH, we therefore
have to specify the full path.

This fixes the missing icon/name of the Mdk window in gnome-shell.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4494>