tests/input-capture: Flush input after client sync
The sync is the client telling it released the capture session at a
specific position, but it's not guaranteed that the release call ended
up updating the state in the input thread. To address this, flush the
input (thus ensuring everything queued to the input thread has been
processed), before asserting the position.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4486>