gpu/kms: Dispose old outputs when replacing
We must disassociate the CRTC and old output when we're replacing
outputs on hotplug. A new association will be set up when applying
the new assignments from the to-be applied configuration. We can't rely
on the last reference to be released until it's too late because
MetaOnscreenNative holds on to it until an async page flip callback.

This also needs a warning about trying to unset a output's monitor if
there wasn't one, because disposed monitor's will try to do this on the
defunct output ref it has before releasing them.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4177
Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4148
Fixes: 5a3e1f9009 ("gpu: Move output/ctrt/modes lifetime management to backend")
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4500>