clutter: Introduce ClutterClickGesture for recognizing clicks
Introduce ClutterClickGesture, a ClutterGesture to recognize clicks
and taps, replacing ClutterTapAction and ClutterClickAction.

It's a subclass of the new ClutterPressGesture and can do all the
things the ClickAction was able to do except for recognizing long-presses.
This is what the ClutterLongPressGesture we'll introduce in a
subsequent commit will be for.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2857>