wayland: Disallow XDND using touch devices
Even though the compositor is able to drive DnD out of touch events,
this is unfortunately doomed to fail in Xwayland, since it's not
possible to actively grab a currently active touch sequence so events
are handled somewhere else. This would be typically possible if using
only pointer emulated events, which we don't know if the client is
using, so it's better to avoid the broken state altogether.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4526>