mdk/session: Init synchronously
We initialized asynchronously in a thread, and that was fine, except
for one detail: we attached the libei source before waiting for the
default seat in the initialization thread. This caused problems because
it meant the main thread was dispatching libei events at the same time
as the initialization thread that waited for the default seat.

We could fix this by changing the order of waiting for the seat and
attaching the source, but lets just initialize things synchronously on
the main thread instead, as in practice it's just a couple of D-Bus
round trips anyway, as otherwise we have a period of time where MdkEi is
created, but not safe to access until the session 'ready' signal is
signalled.

This fixes some sporadic devkit startup failures.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4529>