multi-texture-format: Reintroduce snippet caching
It was originally added in 7c8f1b4cc0 but got lost in 9c38f2dc42.

These snippets are retrieved anew every time a window is resized. But
callers never modify them, they're effectively read-only so cache them
at the place of creation.

This is required to convince the pipeline hash that each reuse of the
same snippet really is the same snippet and so the pipeline is unchanged.
`_cogl_pipeline_snippet_list_equal` only does shallow comparisons and
there's no need right now to reimplement it as a deep comparison.

This eliminates the log message:

> Over 50 separate %s have been generated which is very unusual,
> so something is probably wrong!

Fixes: 9c38f2dc42 ("multi-texture: Support color model transforms ...")
Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4208
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4534>