core: Keep also track of pad existence at MetaTabletActionMapper
This generic parent object became a little bit too oriented to tablet
tools for its cycle_output() builtin action. It should also keep track
of pads and their configuration so that this action keeps working when
assigned to pad buttons.

While at it, correct the MetaPadActionMapper instance field in the
object struct. This should be MetaTabletActionMapper and not GObject,
even though both structs happen to be the same size.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4215
Fixes: 6c77af1493 ("core: Add a MetaTabletActionMapper as parent of the MetaPadActionMapper")
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4538>