clutter: Add ClutterScrollFlags member to scroll ClutterEvents
These flags at the moment contain a single additional modifier, in
order to express that the scrolling direction is inverted, relative
to the physical motion on the device.

This flag is at the moment unused, and will be populated in later
commits.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4552>