tests/a11y-mouse-keys: Use test_run_tests helper to properly skip
The test case uses VKMS, and if the kernel module is not loaded, it will
fail. The meta_context_test_run_tests helper correctly skips the test in
that case, so use that instead of doing something similar in the test
case.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4560>