kms: Configure primary GPU in headless mode if it does't support KMS
If a GPU does not support KMS or if KMS initialization fails for
some reason, try configuring it in headless mode if it is tagged
as the primary GPU (via udev).

Currently, Mutter ignores GPU devices that do not support KMS but
there are situations where this is not desirable. For example,
- Some newer Intel GPUs have display/modeset capability disabled
  (therefore, no KMS support) and are intended to be used as
  render-only devices.

- In some virtualization environments, the primary GPU (e.g SRIOV
  enabled GPU VF) may not support KMS but can still be used for
  rendering.

- Some drivers provide an option to disable the display capability
  at boot time (e.g, disable_display module parameter provided by
  i915 and xe) for various reasons.

Given the above use-cases, Mutter should not assume that a GPU
always supports KMS and disregard it if KMS initialization fails,
especially when it is explicitly tagged as the primary GPU via
udev. Instead, it should configure the GPU in headless mode and
try to proceed.

Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4561>
(cherry picked from commit f40bfa12105d0996db230109cc98ccda27600f54)