color-manager: Build a more unique monitor id
If a physical monitor is connected twice, via different connectors, the
monitor id here would not be unique, and the mapping between color
device and monitor could get wrong.

Always use the connector if available to build the monitor id we use for
the mapping.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4574>