tests/monitors: Do not re-use output serials
This currently has bad consequences, namely it will trip up the color
device which relies on colord naming which does not include the
connector. So if the monitors are the same (excluding the connector), we
get warnings about not being able to find, or existing profiles.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4574>