screen-cast-stream-src: Fix sync_fd leak
The function cogl_context_get_latest_sync_fd() is transfer full, so
the caller must take ownership of the returned fd, and clean it up
after use. Furthermore, drmSyncobjImportSyncFile() is transfer none,
and this is the only place where the syncfd is used.

Change the declaration of the sync_fd variable to be g_autofd, and
let this fd be closed when the function returns.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4251
Fixes: c7105a70f1b156c089660260163c0c5060f3e9ac
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4577>