constraints: Ignore size limits for fullscreen Wayland windows
We don't care that there is a limit, we'll fullscreen anyway, and put a
black backdrop if the client doesn't match the size.

The client having a too large surface is problematic, however, as it'll
"leak" into any neighbouring monitor, unless we start to clamp.

So far only the too-small case is tested, what to do with too large
surfaces is unclear.

Related: https://gitlab.gnome.org/GNOME/mutter/-/issues/4249
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4587>