window-actor: Handle visibility condition for suspending window
We were looking at the MetaWindow::hidden state when deciding whether to
let a window enter the suspended state, but we can replace this with
handling effective visibility in the window actor. This has the benefit
of also covering the case where the screen shield is showing in GNOME
Shell, as the window actor itself will effectively be unmapped while the
screen shield is showing.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4588>