wayland/color-management: Map from WaylandOutput instead of Monitor
The previous hash table mapped from a MetaMonitor to the
MetaWaylandColorManagementOutput. When the MetaMonitor of the
MetaWaylandOutput changed, and the MetaWaylandOutput was later
destroyed, we would not find it in the hash table again and thus not
destroy it.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4262
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4638>
(cherry picked from commit cb6a5b9e6be0b72f0d73d4c091c31b6f7027322d)