constraints: Don't constrain size limits when applying Wayland commits
Size limits are for the window manager to make decisions about how to
manage the window; if the window itself ignores its own size limits, we
shouldn't try to tell it otherwise.

This fixes an issue where LibreOffice would set a size limit on the
splash screen that was smaller than the splash screen it provided
itself, which resulted in us trying to resize it when mapping it.
Resizing it tried to make it use the non-placed rectangle, effectively
basing the size-limit constraining on 0x0+0+0.

A test case is added that imitates the invalid state on map that
LibreOffice provides that caused the splash screen to not be mapped
according to the window management rules (i.e. in the center according
to the current defaults).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4593>