window-drag: Make tear down explicit
Tear down in dispose() and actively destroy (which runs dispose and
unrefs) when the drag should end. This will in a follow up allow keeping
the ended drag instance alive as a basis for calculating window
positions that were affected by the drag.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4607>