window-drag: Don't move unshaken fixed windows
When dragging a maximized or tiled window, they are not actually moved
until they are "shaken loose". We still tried to move them, only that it
didn't have any effect due to the constraints keeping them in place.

This avoids incorrectly setting the unconstrained rectangle of the
dragged window to the maximized window size.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4607>