mutter: Launch child command after first monitor appears
When launched with --devkit, the initial monitor isn't created
immediately, but by the mutter-devkit executable after mutter itself is
created. To have windows created from the executed child command be
launched when there is a monitor, as one would expect, wait with
launching until so is the case.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4600>