backends/output: Use the current state to check if privacy screen is enabled
To check if the privacy screen was enabled we used a cached value but
this was never set and thus the privacy screen was always marked as
unavailable.

Instead get the actual state from the output implementation and use this
value to check the current state of the privacy screen.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4259
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4603>