window/wayland: Get size from resize drags when changing state
When we're starting a resize drag, we're indirectly triggering a
surface-state change configuration when adding the 'resizing'
xdg_toplevel state. The state change handler inherits the last
configuration, and if that configuration was of a 'unmaximize'
configuration without any saved rect (i.e. window mapped as maximized),
the sent size was 0x0.

However, the specification for the 'resizing' xdg_toplevel state is that
the size is a maximum, and thus 0x0 makes little sense. Address this by
deriving the size in the configuration from the active resize grab, if
there is such, avoiding sending 0x0 when adding the 'resizing' state.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4281
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4658>