backends/native/kms-connector: Notify disconnect events after a timeout
Before the recent reverts and since commit 313860e2fab we were waiting
for a delay before advertising any KMS state change.

This allowed to properly address the issue it meant to fix (avoiding
monitors re-configurations on quick monitors re-connections, signal
loss or inputs probing), but it has also the side effect of causing a
late notification for KMS events that we want to address immediately
(such as privacy screen or gamma changes).

Since the change was only meant to delay the connector's connection
state changes and, in particular, the disconnections, we can just handle
this in the KMS connector side.

So now if a disconnect event happens:
 - We start a per-connector timeout and do nothing so that the rest of
   the stack won't notice any connector-related change.
   Then, depending on the future events:
   1) If nothing happens after a delay (3 seconds as before), we make
      the KMS machinery to verify again the current state and eventually
      make it to emit a resource change
   2) If the connector gets re-connected before the delay is over, the
      timeout is cancelled and all the other potential changes are
      checked again (potentially leading to a NO-CHANGE).

In this way, all the other kinds of KMS changes that are affecting other
connectors or other aspects of the same connector won't be delayed and
reported when they happen

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4287
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4610>