backends/native/kms-connector: Do not assume !connected == disconnected
In drm a connection status can be connected, disconnected or unknown.

In this third case, as per the documentation
 > DRM_MODE_UNKNOWNCONNECTION is used when it could be either
   [connected or disconnected].

As per this, do not assume that if a connector is in the unknown state
then it's disconnected, as there is no way we could even try to enable
it otherwise.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4610>