stream-src/virtual: Invalidate EIS viewport when re-enabling
When a virtual monitor stream renegotiates the size, the stream is
disabled, renegotiated, then re-enabled. During renegotiation the size,
thus monitor configuration, will change, but we won't be notified about
the monitors changing, since we don't listen for such changes while the
source is disabled.

To not end up with out of date EIS viewports corresponding to the
stream, invalidate them when re-enabling, so that absolute input events
will work with the new virtual monitor size.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4622>