tests: Raise the timeout on wayland-unit from 1 minute to 5 minutes
The test is huge and on a very fast machine it takes 40 seconds. So in
Debian and Ubuntu builds we're finding a lot of machines/architectures
where it regularly requires more than a minute to complete, but times
out instead and fails the package build. Same as 7b06ec1556.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4624>