tests/wayland-unit: Wait for window to be gone before exiting some tests
The /wayland/toplevel/fixed-size-fullscreen-exceeds test was flaky
because the window from the test before
(/wayland/toplevel/fixed-size-fullscreen) would sometimes still be there
when the second test ran. This caused the initial "wait for window to
appear" statement to return too early, before the actual window
appeared.

Fix this by waiting on the window to disappear before exiting the
previous test. Do the same for this test too, to avoid the same issue if
we would add another test with the same test client after.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4625>