window: Let unplaced unconstrained rect follow the constrained rect monitor
On X11, a window is constructed with an "unconstrained" rectangle that
is kept track of. When a window is mapped as maximized, the constrained
rect is the one of the maximized window, and to not have the
unconstrained rect end up on a different monitor, move it along, so that
when we unmaximize, the originally unconstrained rect isn't used for
determining what monitor to perform the actual placement on.

This fixes an issue where one maps an X11 window maximized on a
secondary monitor, minimizes it, unmaximizes, then unminimizing it,
placed it on the first monitor.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4627>