constraints/placement: Also consider resize flag when checking
When checking if a placement rule constraint is satisfied, the placed
rect was checked against the work area, and if a slide or flip flag was
set, consider it non-satisfied if it extended outside the work area. If
a popup only had the resize flag set, this check considered the
placement of a popup extending outside of the work area as satisfied,
despite the intention being that it should resize to be fully visible.
Fix this by also considering the resize flag.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4628>