window-config: Make caller mark config as initial
The logic to set the is-initial state implicitly when creating a new
config from another one was specific to the X11 window usage. Lets move
this is-initial logic to the caller, so that we can make new_from() more
generally useful.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4673>