clutter: Rename ClutterSprite device getter
Move away from making all ClutterSprites have a ClutterInputDevice,
and only set it up on the sprites that are lead by a single device.
This essentially applies to tablet tools, which "happens" to be the
case for all instances of this function being called.

In exchange, use the virtual source device from the seat, in the
cases where we generate events that do not directly come from user
input.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4644>