clutter: Add CLUTTER_KEY_STATE event type
By convention, ClutterEvents generated have the xkb state
applying at the time of receiving the event, this distinction
is important for keys that do change the active modifier set,
and is at odds with the expectations for wl_keyboard that we
will emit the latest modifier state.

We so far use the ClutterKeymap for that, as querying it will
always contain the latest keyboard state as known by the main
thread, but this is still slightly racy as multiple
modifier-changing key events might have queued up at the time
to query the keymap first.

This event will work as a race-free replacement, as keyboard
state changes will be queued up in the right relative order to
the key press/release events that trigger them, with later
key events maybe being paired with their own state change event.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4655>