clutter/stage: Always flush after painting stage to custom framebuffer
Not doing this ends up with cyclic framebuffer dependencies if painting
the stage ended up depending on a offscreen that itself paints from the stage,
e.g. an offscreen effect, and if we'd end up using the texture when
painting the stage with the target framebuffer as part of the painting.

This happened when using the greyscale effect that is part of the
screen time feature, combined with the zoom effect and pressing print
screen.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8634
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4654>
(cherry picked from commit 3078b5393fdc78b9cd1f3e344202f69f5091dc74)